requirements.txt
pandas==2.3.1
matplotlib==3.10.3
mplfinance==0.12.10b0
requests==2.31.0
websocket-client==1.8.0
TA-Lib==0.6.4
contourpy==1.3.3
cycler==0.12.1
fonttools==4.59.0
kiwisolver==1.4.8
numpy==2.3.2
packaging==25.0
pillow==11.3.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
setuptools==80.9.0
six==1.17.0
tzdata==2025.2
dotenv==0.9.9

